% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctors.R
\name{aardvark_state_corrections}
\alias{aardvark_state_corrections}
\title{Corrections function for aardvark}
\usage{
aardvark_state_corrections(signals_list, ...)
}
\arguments{
\item{signals_list}{list of signals as returned from `covidcast_signals()`
uses the list (no aggregation)}

\item{...}{named corrections parameters passed to `default_state_params()`}
}
\value{
a list of signals the same length and format as the input with
  updated entries in the `value` column

  If the parameter `corrections_db_path` gives a path, intermediate output
  is written to that sqlite database
}
\description{
Corrections function for aardvark
}
