% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetters.R
\name{county_subsetter}
\alias{county_subsetter}
\title{Only applies corrections to some county data}
\usage{
county_subsetter(x, params)
}
\arguments{
\item{x}{a data frame as returned by `evalcast:::download_signal()`}

\item{params}{a list of parameters, typically `correction_pars`}
}
\value{
a subset of x
}
\description{
Only applies corrections to some county data
}
