% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrector-zyzzyva.R
\name{default_county_params}
\alias{default_county_params}
\title{Default parameters for  county case corrections}
\usage{
default_county_params(
  data_source = "usa-facts",
  signal = c("confirmed_incidence_num"),
  geo_type = "county",
  window_size = 14,
  backfill_lag = 30,
  excess_cut = 0,
  size_cut = 20,
  sig_cut = 3,
  sig_consec = 2.25,
  time_value_flag_date = Sys.Date() + 1,
  multinomial_preprocessor = TRUE
)
}
\arguments{
\item{data_source}{covidcast data_source for corrections}

\item{signal}{vector of covidcast signals to correct}

\item{geo_type}{the geo_type we're forecasting}

\item{window_size}{size of rolling window}

\item{backfill_lag}{how far back do we fill the spikes}

\item{excess_cut}{currently ignored}

\item{size_cut}{"outliers" are ignored if they are smaller in magnitude}

\item{sig_cut}{t-statistic cut off for marking outliers}

\item{sig_consec}{slightly smaller t-statistic if consecutive}

\item{time_value_flag_date}{no corrections after this date}

\item{multinomial_preprocessor}{logical, do we run the preprocessor}
}
\value{
A list of parameter values
}
\description{
Default parameters for  county case corrections
}
\examples{
default_state_params(window_size=21)
}
