% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctors.R
\name{default_state_params}
\alias{default_state_params}
\title{Default parameters for aardvark state death corrections}
\usage{
default_state_params(
  signals_to_correct = "response",
  window_size = 14,
  backfill_lag = 30,
  excess_cut = 0,
  size_cut = 20,
  sig_cut = 3,
  sig_consec = 2.25,
  time_value_flag_date = Sys.Date() + 1,
  multinomial_preprocessor = TRUE,
  corrections_db_path = NULL,
  integer_tol = 1e-06,
  ...
)
}
\arguments{
\item{signals_to_correct}{either "response" or "all"}

\item{window_size}{size of rolling window}

\item{backfill_lag}{how far back do we fill the spikes}

\item{excess_cut}{currently ignored}

\item{size_cut}{"outliers" are ignored if they are smaller in magnitude}

\item{sig_cut}{t-statistic cut off for marking outliers}

\item{sig_consec}{slightly smaller t-statistic if consecutive}

\item{time_value_flag_date}{no corrections after this date}

\item{multinomial_preprocessor}{logical, do we run the preporcessor}

\item{corrections_db_path}{path to database if storing results}

\item{integer_tol}{small number to handle integer checks}

\item{...}{ignored}
}
\value{
A list of paramter values
}
\description{
Default parameters for aardvark state death corrections
}
\examples{
default_state_params(window_size=21)
}
