% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz-apply_corrections.R
\name{get_data_corrections}
\alias{get_data_corrections}
\title{Retrieve a tibble of data corrections from the corrections database}
\usage{
get_data_corrections(db_path, geo_type)
}
\arguments{
\item{db_path}{the path for the SQLite database, for example
`~/data_corrections.sqlite`}

\item{geo_type}{the `geo_type`, one of `"state"` or `"county"`}
}
\value{
a tibble of corrections data
}
\description{
Reads all records in the corrections database and returns a tibble
with columns noted above
}
