% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_target_period.R
\name{get_target_period}
\alias{get_target_period}
\title{Gets the forecast target period}
\usage{
get_target_period(forecast_date, incidence_period, ahead)
}
\arguments{
\item{forecast_date}{can be a vector of dates}

\item{incidence_period}{one of "epiweek" or "day"}

\item{ahead}{how many epiweeks/days ahead are you forecasting?}
}
\value{
a tibble with variables "start" and "end"
}
\description{
This function gives the start and end dates of the target period,
based on the system described in the COVIDHub rules here:
https://github.com/reichlab/covid19-forecast-hub/blob/master/data-processed/README.md
}
\details{
For week-ahead forecasts with forecast_date of Sunday or Monday of EW12, a
1 week ahead forecast corresponds to EW12 and should have target_end_date of
the Saturday of EW12. For week-ahead forecasts with forecast_date of Tuesday
through Saturday of EW12, a 1 week ahead forecast corresponds to EW13 and
should have target_end_date of the Saturday of EW13.
}
