% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_corrections.R
\name{identical_schema}
\alias{identical_schema}
\title{A simple check to ensure two data frames have same column names and classes}
\usage{
identical_schema(x, y)
}
\arguments{
\item{x}{the first df}

\item{y}{the second df}
}
\value{
a boolean `TRUE` or `FALSE`
}
\description{
A simple check to ensure two data frames have same column names and classes
}
