% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrector-states.R
\name{make_state_corrector}
\alias{make_state_corrector}
\alias{make_aardvark_corrector}
\title{Corrector for state forecasts}
\usage{
make_state_corrector(
  params = default_state_params(),
  corrections_db_path = NULL,
  dump_locations = c("as", "gu", "mp", "vi"),
  manual_flags = NULL
)

make_aardvark_corrector(
  params = default_state_params(),
  corrections_db_path = NULL,
  dump_locations = c("as", "gu", "mp", "vi"),
  manual_flags = NULL
)
}
\arguments{
\item{params}{a tibble with corrections parameters. The number of
rows is the number of signals used by the forecaster. The
tibble is most easily generated with [default_state_params()].}

\item{corrections_db_path}{path to store results, NULL by default}

\item{dump_locations}{character vector of locations to ignore}

\item{manual_flags}{tibble with five columns: `geo_value`, `time_value`,
`data_source`, `signal`, and `max_lag` indicating combinations of
of dates, locations, etc to manually flag and back distribute. `max_lag`
determines how far to backfill. For multiple times at one location,
`time_value` may be a list column.}
}
\value{
A function that takes a list of covidcast signals as the only
  argument
}
\description{
This function produces another function to create corrections. It expects
the signals from covidcast to be a list (or a single signal).
}
\section{Functions}{
\itemize{
\item \code{make_aardvark_corrector}: alias to avoid destroying production
}}

\examples{
make_state_corrector(default_state_params(window_size=21))
}
