% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrector-zyzzyva.R
\name{make_zyzzyva_corrector}
\alias{make_zyzzyva_corrector}
\title{Corrector for county forecasts}
\usage{
make_zyzzyva_corrector(
  params = default_county_params(),
  corrections_db_path = NULL,
  dump_locations = NULL
)
}
\arguments{
\item{params}{a tibble with corrections parameters. The number of
rows is the number of signals used by the forecaster. The
tibble is most easily generated with [default_state_params()].}

\item{corrections_db_path}{path to store results, NULL by default}

\item{dump_locations}{character vector of locations to ignore}
}
\value{
A function that takes a list of covidcast signals as the only
  argument
}
\description{
This function produces another function to create corrections. It expects
the signals from covidcast to be a list (or a single signal).
}
\examples{
make_zyzzyva_corrector(default_state_params(window_size=21))
}
