% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctors.R
\name{state_corrector}
\alias{state_corrector}
\title{Creates additional columns and flags outliers for state-level data}
\usage{
state_corrector(x, params)
}
\arguments{
\item{x}{a data frame as returned by `evalcast:::download_signal()`}

\item{params}{a list of parameters, typically `correction_pars`}
}
\value{

}
\description{
Creates additional columns and flags outliers for state-level data
}
