% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrections_utils.R
\name{unchop_cartesian_mapping}
\alias{unchop_cartesian_mapping}
\title{Form and combine Cartesian products of mappings encoded by rows of a data frame}
\usage{
unchop_cartesian_mapping(df, ...)
}
\arguments{
\item{df}{a data frame encoding a multivariate mapping in a condensed form}

\item{...}{a partition of the columns of the data frame into groups; each
argument should either be a tidy-select argument indicating key columns in
a group with no val columns, or a formula of the form <keyselector> ~
<valselector>, where <keyselector> and <valselector> are tidy-select
arguments}
}
\value{
a data frame conceptually formed by this process: (1) every non-list
  column is converted to a list column with entries of length 1; (2) within
  each group of columns, each row should have entries of length 1 or the same
  non-1 number --- the former are \code{rep}'d to match the length of the
  latter; (3) for each row, take the Cartesian product between data frames
  formed from the column groups; (4) `bind_rows` to combine the Cartesian
  products together. Or alternatively, by simply calling \code{unchop} for
  each column group separately.
}
\description{
Form and combine Cartesian products of mappings encoded by rows of a data frame
}
\examples{

tbl1 <- tibble::tribble(
  ~location, ~time, ~unique_time_description,
  c("AK","AL"), c(as.Date("2021-01-01"), as.Date("2021-01-02"), as.Date("2021-01-03")), c("New Year's Day","Other","Other"),
  "AK", as.Date("2021-03-29"), "Seward's Day"
)
zookeeper:::unchop_cartesian_mapping(
  tbl1,
  location, time ~ unique_time_description
)

## In the above example, there must not be duplicate (location, time) values
## in the result. In the below examples, duplicate (location, time) values
## are allowed, but entirely duplicated rows are not.

tbl2 <- tibble::tribble(
  ~location, ~time, ~time_tag,
  c("AK","AL"), as.Date("2021-03-29"), "Not New Year's",
  "AK", as.Date("2021-03-29"), "Seward's Day"
)
zookeeper:::unchop_cartesian_mapping(
  tbl2,
  location, starts_with("time")
)

tbl3 <- tibble::tribble(
  ~time, ~location, ~tag,
  as.Date("2021-01-01"), c("AK","AL"), "New Year's Day",
  c(as.Date("2021-01-02"),as.Date("2021-01-03")), c("AK","AL"), "Other",
  as.Date("2021-03-29"), c("AK","AL"), c("Seward's Day","Other")
)
zookeeper:::unchop_cartesian_mapping(
  tbl3,
  time, location ~ tag
)

}
