% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zookeeper_corrections.R
\name{zookeeper_corrections}
\alias{zookeeper_corrections}
\title{Apply corrections to case/death data prior to forecasting}
\usage{
zookeeper_corrections(df)
}
\arguments{
\item{df}{}
}
\value{
a data frame with additional columns
}
\description{
Apply corrections to case/death data prior to forecasting
}
