% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reichlab-formatter.R
\name{format_predictions_cards_for_reichlab_submission}
\alias{format_predictions_cards_for_reichlab_submission}
\title{Format evalcast predictions_cards for submission to Reich Lab}
\usage{
format_predictions_cards_for_reichlab_submission(predictions_cards)
}
\arguments{
\item{predictions_cards}{list of predictions cards of the type returned by evalcast}
}
\value{
a tibble for conversion to csv
}
\description{
Format evalcast predictions_cards for submission to Reich Lab
}
\examples{
baby_predictions = get_predictions(
  baseline_forecaster, "baby",
  tibble::tibble(
    data_source=c("jhu-csse", "usa-facts"),
    signal = c("deaths_incidence_num","confirmed_incidence_num"),
    start_day=lubridate::ymd("2020-09-15")),
  lubridate::ymd("2020-10-01"),
  "epiweek", ahead = 1:4, "state", "mi")
long_tbl = format_predictions_cards_for_reichlab_submission(baby_predictions)
}
