% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reichlab-formatter.R
\name{format_predictions_for_reichlab_submission}
\alias{format_predictions_for_reichlab_submission}
\title{Format evalcast predictions_cards for submission to Reich Lab}
\usage{
format_predictions_for_reichlab_submission(
  predictions_cards,
  geo_values_to_filter = NULL
)
}
\arguments{
\item{predictions_cards}{data frame of predictions of the type returned by
evalcast. You must pass in a single df.}

\item{geo_values_to_filter}{character string or list thereof of geo_values to exclude from the
final output.}
}
\value{
a tibble for conversion to csv
}
\description{
Each forecaster produces a data frame of predictions. It is easiest to just
`bind_rows()` and pass the result in to this function.
}
