% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrector-aardvark.R
\name{make_aardvark_corrector}
\alias{make_aardvark_corrector}
\title{Corrector for state forecasts}
\usage{
make_aardvark_corrector(
  params = default_state_params(),
  corrections_db_path = NULL,
  dump_locations = c("as", "gu", "mp", "vi")
)
}
\arguments{
\item{params}{a tibble with corrections parameters. The number of
rows is the number of signals used by the forecaster. The
tibble is most easily generated with [default_state_params()].}

\item{corrections_db_path}{path to store results, NULL by default}

\item{dump_locations}{character vector of locations to ignore}
}
\value{
A function that takes a list of covidcast signals as the only
  argument
}
\description{
This function produces another function to create corrections. It expects
the signals from covidcast to be a list (or a single signal).
}
\examples{
make_aardvark_corrector(default_state_params(window_size=21))
}
