% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrections_utils.R
\name{make_manual_corrections}
\alias{make_manual_corrections}
\title{(2) This function makes corrections on manually flagged columns
This function will return the original dataframe with two additional columns:
`corrected` and `special_flag`, in which the latter one is used to avoid duplicate corrections}
\usage{
make_manual_corrections(df, special_flags)
}
\description{
(2) This function makes corrections on manually flagged columns
This function will return the original dataframe with two additional columns:
`corrected` and `special_flag`, in which the latter one is used to avoid duplicate corrections
}
