% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrections_utils.R
\name{make_manual_flags}
\alias{make_manual_flags}
\title{Two helper functions
 (1) Add manually flagged columns, this function will create columns like
 "flag_bad_va", "flag_bad_wa", etc.}
\usage{
make_manual_flags(df, special_flags)
}
\description{
Two helper functions
 (1) Add manually flagged columns, this function will create columns like
 "flag_bad_va", "flag_bad_wa", etc.
}
