% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial_preprocessor.R
\name{multinomial_preprocessor}
\alias{multinomial_preprocessor}
\alias{multinomial_preprocesser}
\title{Preprocesses the response values in a data frame.}
\usage{
multinomial_preprocessor(dat, response, max_lag = Inf)

multinomial_preprocesser(dat, response, max_lag = Inf)
}
\arguments{
\item{dat}{a data frame with columns reference_date, location, variable_name, and value
(and possibly some other columns we will ignore.)}

\item{response}{the name of the response variable.}

\item{max_lag}{maximum amount counts can be moved. Default to `Inf`}
}
\value{
a data frame of the same format as `dat`
}
\description{
Replaces zeroes (which we suspect are due to lagged reporting) by samples from a multinomial.
}
\section{Functions}{
\itemize{
\item \code{multinomial_preprocesser}: an alias for previously misspelled function
}}

