% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{rmultinom_wrapper}
\alias{rmultinom_wrapper}
\title{Wrapper for \link[stats:Multinom]{stats::rmultinom}}
\usage{
rmultinom_wrapper(n, size, prob)
}
\description{
The \link{with_test} functions are restricted due to potential issues with bytecode and/or JIT compilation of base/external packages and/or cleanliness.  The \link{stub} function approach has issues due to the complexity of code and environment transformation.
}
