% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz-apply_corrections.R
\name{update_corrections}
\alias{update_corrections}
\title{Replace corrections data for a geo type with new data}
\usage{
update_corrections(db_path, geo_type, new_df)
}
\arguments{
\item{db_path}{the path for the SQLite database, for example
`~/corrections.sqlite`}

\item{geo_type}{the `geo_type`, one of `"state"` or `"county"`}

\item{new_df}{the new data to replace the old}
}
\value{
`TRUE` invisibly
}
\description{
Overwrites the corrections records for a geo type with new data after checking that schemas match.
}
